Add default impementations for negotiator trait
